requirements.txt
# Requirements for Heart Disease LLM Research

# Core Data Science
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# LLM APIs
openai>=1.0.0
google-generativeai>=0.3.0
dashscope>=1.14.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.0.0

# Optional (for advanced visualizations)
plotly>=5.14.0
